Readme
The program was created using python in a virtual pipenv. 
Scrapy was installed and used for scraping the names and prices of the vehicles.
By reading the source of the website from which data was collected using inspect, the required names and prices of the vehicles were obtained.
The obtained data was duly coded using python.
The program was executed in pycharm.